Move gameinterface to gamelogic package
